requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10

# Vector database & ML
faiss-cpu==1.7.4
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.36.2

# AWS
boto3==1.34.0
mangum==0.17.0

# Database
pymongo==4.6.0
psycopg2-binary==2.9.9

# RAG & LLM
langchain==0.1.0
openai==1.3.7

# Frontend (if using Streamlit)
streamlit==1.29.0

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
click==8.1.7
